Feat(bidList): Change bidQuantity Type(BigDecimal to Double)

+++ src/main/java/com/nnk/springboot/dto/bid/BidListDto.java
@@ -2,13 +2,11 @@ package com.nnk.springboot.dto.bid;
 
 import lombok.*;
 
-import java.math.BigDecimal;
-
 @Data
 @Builder
 public class BidListDto {
     private Integer id;
     private String account;
     private String type;
-    private BigDecimal bidQuantity;
+    private Double bidQuantity;
 }
